+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.2] - 2025-2-15
 
+### Added
+
+- Added debug output
+
 ### Fixed
 
 - Fixed custom block explosion